e2e/cli: code review comments, restructing and cleanup
